+++ lib/install.ps1
@@ -1,6 +1,3 @@
-. "$PSScriptRoot\autoupdate.ps1"
-. "$PSScriptRoot\buckets.ps1"
-
 function nightly_version($date, $quiet = $false) {
     $date_str = $date.tostring("yyyyMMdd")
     if (!$quiet) {
@@ -10,8 +7,7 @@ function nightly_version($date, $quiet = $false) {
 }
 
 function install_app($app, $architecture, $global, $suggested, $use_cache = $true, $check_hash = $true) {
-    $app, $bucket, $null = parse_app $app
-    $app, $manifest, $bucket, $url = Find-Manifest $app $bucket
+    $app, $manifest, $bucket, $url = Get-Manifest $app
 
     if(!$manifest) {
         abort "Couldn't find manifest for '$app'$(if($url) { " at the URL $url" })."
@@ -54,7 +50,8 @@ function install_app($app, $architecture, $global, $suggested, $use_cache = $tru
     $persist_dir = persistdir $app $global
 
     $fname = dl_urls $app $version $manifest $bucket $architecture $dir $use_cache $check_hash
-    pre_install $manifest $architecture
+    Invoke-HookScript -HookType 'pre_install' -Manifest $manifest -Arch $architecture
+
     run_installer $fname $manifest $architecture $dir $global
     ensure_install_dir_not_in_path $dir $global
     $dir = link_current $dir
@@ -68,7 +65,7 @@ function install_app($app, $architecture, $global, $suggested, $use_cache = $tru
     persist_data $manifest $original_dir $persist_dir
     persist_permission $manifest $global
 
-    post_install $manifest $architecture
+    Invoke-HookScript -HookType 'post_install' -Manifest $manifest -Arch $architecture
 
     # save info for uninstall
     save_installed_manifest $app $bucket $dir $url
@@ -81,38 +78,6 @@ function install_app($app, $architecture, $global, $suggested, $use_cache = $tru
     success "'$app' ($version) was installed successfully!"
 
     show_notes $manifest $dir $original_dir $persist_dir
-}
-
-function locate($app, $bucket) {
-    Show-DeprecatedWarning $MyInvocation 'Find-Manifest'
-    return Find-Manifest $app $bucket
-}
-
-function Find-Manifest($app, $bucket) {
-    $manifest, $url = $null, $null
-
-    # check if app is a URL or UNC path
-    if($app -match '^(ht|f)tps?://|\\\\') {
-        $url = $app
-        $app = appname_from_url $url
-        $manifest = url_manifest $url
-    } else {
-        # check buckets
-        $manifest, $bucket = find_manifest $app $bucket
-
-        if(!$manifest) {
-            # couldn't find app in buckets: check if it's a local path
-            $path = $app
-            if(!$path.endswith('.json')) { $path += '.json' }
-            if(test-path $path) {
-                $url = "$(resolve-path $path)"
-                $app = appname_from_url $url
-                $manifest, $bucket = url_manifest $url
-            }
-        }
-    }
-
-    return $app, $manifest, $bucket, $url
 }
 
 function dl_with_cache($app, $version, $url, $to, $cookies = $null, $use_cache = $true) {
@@ -304,7 +269,7 @@ function dl_with_cache_aria2($app, $version, $manifest, $architecture, $dir, $co
         $oriConsoleEncoding = [Console]::OutputEncoding
         [Console]::OutputEncoding = New-Object System.Text.UTF8Encoding
 
-        Invoke-Expression $aria2 | ForEach-Object {
+        Invoke-Command ([scriptblock]::Create($aria2)) | ForEach-Object {
             # Skip blank lines
             if ([String]::IsNullOrWhiteSpace($_)) { return }
 
@@ -387,15 +352,25 @@ function dl_with_cache_aria2($app, $version, $manifest, $architecture, $dir, $co
 
 # download with filesize and progress indicator
 function dl($url, $to, $cookies, $progress) {
-    $reqUrl = ($url -split "#")[0]
-    $wreq = [net.webrequest]::create($reqUrl)
-    if($wreq -is [net.httpwebrequest]) {
-        $wreq.useragent = Get-UserAgent
-        if (-not ($url -imatch "sourceforge\.net" -or $url -imatch "portableapps\.com")) {
-            $wreq.referer = strip_filename $url
+    $reqUrl = ($url -split '#')[0]
+    $wreq = [Net.WebRequest]::Create($reqUrl)
+    if ($wreq -is [Net.HttpWebRequest]) {
+        $wreq.UserAgent = Get-UserAgent
+        if (-not ($url -match 'sourceforge\.net' -or $url -match 'portableapps\.com')) {
+            $wreq.Referer = strip_filename $url
         }
-        if($cookies) {
-            $wreq.headers.add('Cookie', (cookie_header $cookies))
+        if ($url -match 'api\.github\.com/repos') {
+            $wreq.Accept = 'application/octet-stream'
+            $wreq.Headers['Authorization'] = "token $(Get-GitHubToken)"
+        }
+        if ($cookies) {
+            $wreq.Headers.Add('Cookie', (cookie_header $cookies))
+        }
+
+        get_config 'private_hosts' | Where-Object { $_ -ne $null -and $url -match $_.match } | ForEach-Object {
+            (ConvertFrom-StringData -StringData $_.Headers).GetEnumerator() | ForEach-Object {
+                $wreq.Headers[$_.Key] = $_.Value
+            }
         }
     }
 
@@ -743,7 +718,7 @@ function run_installer($fname, $manifest, $architecture, $dir, $global) {
     $installer = installer $manifest $architecture
     if($installer.script) {
         write-output "Running installer script..."
-        Invoke-Expression (@($installer.script) -join "`r`n")
+        Invoke-Command ([scriptblock]::Create($installer.script -join "`r`n"))
         return
     }
 
@@ -823,7 +798,7 @@ function run_uninstaller($manifest, $architecture, $dir) {
     $version = $manifest.version
     if($uninstaller.script) {
         write-output "Running uninstaller script..."
-        Invoke-Expression (@($uninstaller.script) -join "`r`n")
+        Invoke-Command ([scriptblock]::Create($uninstaller.script -join "`r`n"))
         return
     }
 
@@ -946,7 +921,7 @@ function link_current($versiondir) {
         Remove-Item $currentdir -Recurse -Force -ErrorAction Stop
     }
 
-    New-Item -Path $currentdir -ItemType Junction -Value $versiondir | Out-Null
+    New-DirectoryJunction $currentdir $versiondir | Out-Null
     attrib $currentdir +R /L
     return $currentdir
 }
@@ -1056,19 +1031,25 @@ function env_rm($manifest, $global, $arch) {
     }
 }
 
-function pre_install($manifest, $arch) {
-    $pre_install = arch_specific 'pre_install' $manifest $arch
-    if($pre_install) {
-        write-output "Running pre-install script..."
-        Invoke-Expression (@($pre_install) -join "`r`n")
-    }
-}
+function Invoke-HookScript {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('pre_install', 'post_install',
+                     'pre_uninstall', 'post_uninstall')]
+        [String] $HookType,
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [PSCustomObject] $Manifest,
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('32bit', '64bit')]
+        [String] $Arch
+    )
 
-function post_install($manifest, $arch) {
-    $post_install = arch_specific 'post_install' $manifest $arch
-    if($post_install) {
-        write-output "Running post-install script..."
-        Invoke-Expression (@($post_install) -join "`r`n")
+    $script = arch_specific $HookType $Manifest $Arch
+    if ($script) {
+        Write-Output "Running $HookType script..."
+        Invoke-Command ([scriptblock]::Create($script -join "`r`n"))
     }
 }
 
@@ -1197,7 +1178,7 @@ function persist_data($manifest, $original_dir, $persist_dir) {
             # create link
             if (is_directory $target) {
                 # target is a directory, create junction
-                New-Item -Path $source -ItemType Junction -Value $target | Out-Null
+                New-DirectoryJunction $source $target | Out-Null
                 attrib $source +R /L
             } else {
                 # target is a file, create hard link
@@ -1258,5 +1239,16 @@ function test_running_process($app, $global) {
         }
     } else {
         return $false
+    }
+}
+
+# wrapper function to create junction links
+# Required to handle docker/for-win#12240
+function New-DirectoryJunction($source, $target) {
+    # test if this script is being executed inside a docker container
+    if (Get-Service -Name cexecsvc -ErrorAction SilentlyContinue) {
+        cmd.exe /d /c "mklink /j `"$source`" `"$target`""
+    } else {
+        New-Item -Path $source -ItemType Junction -Value $target
     }
 }