+++ libexec/scoop-help.ps1
@@ -2,12 +2,6 @@
 # Summary: Show help for a command
 param($cmd)
 
-. "$PSScriptRoot\..\lib\core.ps1"
-. "$PSScriptRoot\..\lib\commands.ps1"
-. "$PSScriptRoot\..\lib\help.ps1"
-
-reset_aliases
-
 function print_help($cmd) {
     $file = Get-Content (command_path $cmd) -raw
 
@@ -47,4 +41,3 @@ Some useful commands are:"
 }
 
 exit 0
-